requirements.txt
python-binance
pandas
ta
requests
python-telegram-bot